Fix rain rate numerical anomaly
If we get a very very small dr ... one that isn't physically possible due to the tipping bucket,
we can double-count a tip for the purposes of calculating rain rate resulting in a fake spike in rain rate.
Fix is to threshold dr above 1E-5. Also do not log very small rain rates in the database (<1E-4 mm/hr).